linguabridge/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
openai==1.3.0
anthropic==0.7.7
google-generativeai==0.3.2
speechrecognition==3.10.1
pydub==0.25.1
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.34.0
drf-yasg==1.21.7
django-filter==23.5
drf-spectacular==0.26.5
deep-translator==1.1.1
urduhack
pyarabic==0.6.15